refactor(octopus): move CDN edge-block detection into utils
is_edge_block_body, the CDN markers and the token-mint backoff schedule are not
Octopus-specific: every Kraken-based provider (Octopus, EDF, E.ON) mints its JWT
through the same CDN-fronted endpoint and can be blocked the same way. Move them
to utils so the Kraken auth mixin can use them too.

No behaviour change. EDGE_BLOCK_REFRESH_AFTER stays in octopus.py - it is policy
for that component's query path, not a shared helper.
